+++ src/controllers/products/product.controller.ts
@@ -15,9 +15,21 @@ class ProductController {
       createResponse(200, products, "Successfully fetched products"),
     );
   }
+
   async create(req: Request, res: Response) {
     const user = req.user as PayloadUserWithID;
     const result = createProductSchema.parse(req.body);
+
+    const existing = await Product.findOne({ name: result.name });
+
+    if (existing) {
+      const img = new OptimisedImage();
+      existing.imageLinks.forEach((link) => {
+        return img.deleteImageByLink(link);
+      });
+
+      throw new CustomError("Product already exists", 400);
+    }
 
     const product = await Product.create({
       name: result.name,
@@ -41,24 +53,38 @@ class ProductController {
     const product = await Product.findById(result._id);
 
     if (!product) {
+      if (result.imageLinks && result.imageLinks.length > 0) {
+        const img = new OptimisedImage();
+        result.imageLinks.forEach((link) => {
+          return img.deleteImageByLink(link);
+        });
+      }
+
       throw new CustomError("Product not found", 404);
     }
 
     if (result.imageLinks && result.imageLinks.length > 0) {
+      const img = new OptimisedImage();
       product.imageLinks.forEach((link) => {
-        return new OptimisedImage().deleteImageByLink(link);
+        return img.deleteImageByLink(link);
       });
     }
 
-    const updatedProduct = await Product.findByIdAndUpdate(result._id, {
-      name: result.name,
-      desc: result.desc,
-      price: result.price,
-      stock: result.stock,
-      updatedBy: user._id,
-      category: result.category,
-      imageLinks: result.imageLinks,
-    });
+    const updatedProduct = await Product.findByIdAndUpdate(
+      result._id,
+      {
+        name: result.name,
+        desc: result.desc,
+        price: result.price,
+        stock: result.stock,
+        updatedBy: user._id,
+        category: result.category,
+        imageLinks: result.imageLinks,
+      },
+      {
+        new: true,
+      },
+    );
 
     return res.json(
       createResponse(200, updatedProduct, "Successfully updated product"),
@@ -66,6 +92,16 @@ class ProductController {
   }
   async delete(req: Request, res: Response) {
     const result = idSchema.parse(req.body);
+
+    const product = await Product.findOne({ _id: result._id });
+
+    if (!product) {
+      throw new CustomError("Product not found", 404);
+    }
+
+    const img = new OptimisedImage();
+    product.imageLinks.forEach((link) => img.deleteImageByLink(link));
+
     const deletedProduct = await Product.findByIdAndDelete(result._id);
 
     return res.json(
